Add a marker file to a project

set_here() creates an empty file named .here, by default in the current directory. When here encounters such a file, it uses the directory that contains this file as root. This is useful if none of the default criteria apply. You need to restart the R session so that here() picks up the newly created file.

set_here(path = ".", verbose = TRUE)

Arguments

path
[character(1)]
Directory where to create .here file, defaults to the current directory.

verbose
[logical(1)]
Verbose output, defaults to TRUE.